ADDED: order confirmation is now working, order unique constraint is …

+++ spring-boot/initial/src/main/java/com/example/accessingdatamysql/models/CartItem.java
@@ -75,7 +75,7 @@ public class CartItem {
     public Integer getAmount() {
         return amount;
     }
-    public Integer getPrice()
+    public Double getPrice()
     {
         return this.sells.getPrice();
     }
